Fix device type col name, finish controller + gem
You can't use "type" as a column name, which is a bummer, but I guess
I wasn't being as verbose as I ought to have been so really it's a good
thing.

The devices controller only lets you create/destroy subscriptions at
this point.

Also this adds the gem for push notifications, pushmeup.